Grab DuckDB during the first run if it does not exist

+++ get.sh
@@ -4,6 +4,9 @@ set -e
 set -o pipefail
 
 export DUCKDB_VERSION=v0.2.4
-wget https://github.com/cwida/duckdb/releases/download/${DUCKDB_VERSION}/duckdb_cli-linux-amd64.zip -O duckdb_cli-linux-amd64.zip
-unzip -o duckdb_cli-linux-amd64.zip
-rm duckdb_cli-linux-amd64.zip
+
+if [ ! -f duckdb ]; then
+    wget https://github.com/cwida/duckdb/releases/download/${DUCKDB_VERSION}/duckdb_cli-linux-amd64.zip -O duckdb_cli-linux-amd64.zip
+    unzip -o duckdb_cli-linux-amd64.zip
+    rm duckdb_cli-linux-amd64.zip
+fi
